added some repositories
