+++ add.php
@@ -9,7 +9,7 @@ if (!$user->is_auth_user()) {
 
 date_default_timezone_set('Europe/Moscow');
 
-$categories = $query_result->get_all_categories();
+$categories = $categories_queries->get_all_categories();
 $options = ['Выберите категорию'];
 
 foreach($categories as $category) {
@@ -26,10 +26,16 @@ $lot_date_post = check_date('lot-date');
 $validate_form = checkLotForm([$title_post, $category_post, $message_post, $user_file_post, $lot_rate_post, $lot_step_post, $lot_date_post]);
 
 if (isset($_POST['submit']) && !$validate_form) {
-    $lot_category = $query_result->get_categoryId_by_name($category_post['value']);
-    $author_id = $query_result->get_user_by_email($_SESSION['email'])[0]['id'];
-    $new_lot_id = $query_result->add_new_lot($lot_category, $author_id, $_POST['lot-name'], $_POST['message'], $user_file_post['url'], $_POST['lot-rate'], date('Y-m-d H:i:s' ,strtotime($_POST['lot-date'])), $_POST['lot-step']);
-    
+    $lot_category = $categories_queries->get_categoryId_by_name($category_post['value']);
+    $author_id = $user->get_user_data()['id'];
+    $title = $title_post['value'];
+    $description = $message_post['value'];
+    $image = $user_file_post['url'];
+    $start_price = $lot_rate_post['value'];
+    $expire = date('Y-m-d H:i:s' ,strtotime($lot_date_post['value']));
+    $step = $lot_step_post['value'];
+    $new_lot_id = $lots_queries->add_new_lot($lot_category, $author_id, $title, $description, $image, $start_price, $expire, $step);
+
     send_header("Location: http://yeticave/lot.php?lot_id=$new_lot_id");
 }
 ?>
