+++ login.php
@@ -2,14 +2,14 @@
 
 require_once 'init.php';
 
-$categories = $query_result->get_all_categories();
+$categories = $categories_queries->get_all_categories();
 
 $email_post = check_email('email');
 $pass_post = checkTextInput('password');
 $validate_form = checkLotForm([$email_post, $pass_post]);
 
 if (isset($_POST['submit']) && !$validate_form) {
-    $user->auth_user($query_result, $_POST['email'], $_POST['password']);
+    $user->auth_user($users_queries, $_POST['email'], $_POST['password']);
 
     if ($user->is_auth_user()) {
         header('Location: /index.php');
